(vt) Move avro codec, drop spark dep - closes #1664
- The Avro codec for the `VectorTile` type lived in `vectortile`, and
depended on the spark subproject. Doing this forced us to also depend
directly on `spark-core`, or else `vectortile`'s tests wouldn't compile,
docs wouldn't build, and assembly wouldn't run. This commits moves all
sparky code into the `spark` subproject to avoid these problems.
