+++ spark/src/main/scala/geotrellis/spark/io/avro/codecs/Implicits.scala
@@ -8,6 +8,7 @@ object Implicits extends Implicits
 trait Implicits
     extends TileCodecs
     with TileFeatureCodec
+    with VectorTileCodec
     with ExtentCodec
     with KeyCodecs {
   implicit def tileUnionCodec = new AvroUnionCodec[Tile](
