+++ vectortile/build.sbt
@@ -4,6 +4,5 @@ name := "geotrellis-vectortile"
 
 libraryDependencies ++= Seq(
   scalatest % "test",
-  "org.apache.spark" %% "spark-core" % "1.6.2",
   "com.trueaccord.scalapb" %% "scalapb-runtime" % "0.5.32"
 )
